+++ frontend/src/api/courses/index.ts
@@ -1,9 +1,9 @@
 export { default as useApplyCourse } from './applyCourse';
 export { default as useDeleteCourse } from './deleteCourse';
 export { default as useGetCourseInfo } from './getCourseInfo';
-export { default as searchAllCourses } from './searchAllCourses';
 export { default as useGetPaginatedCourses } from './getPaginatedCourses';
 export { default as useGetCourseMaterials } from './getCourseMaterials';
 export { default as useGetClientCourseAndMaterials } from './getClientCourseAndMaterials';
 export { default as useGetAvailableCourses } from './getAvailableCourses';
 export { default as useGetCoursesMap } from './getCoursesMap';
+export { default as useSearchAllCourses } from './searchAllCourses';
