+++ frontend/src/components/CustomNavLink/CustomNavLink.tsx
@@ -6,18 +6,16 @@ import { ICustomNavLinkProps } from './types';
 const CustomNavLink = (
   { children, activeClassName, inactiveClassName, defaultClassName, ...props }: ICustomNavLinkProps,
   ref: ForwardedRef<HTMLAnchorElement>,
-) => {
-  return (
-    <NavLink
-      ref={ref}
-      className={({ isActive }) =>
-        ` ${defaultClassName} ${isActive ? activeClassName : inactiveClassName}`
-      }
-      {...props}
-    >
-      {children}
-    </NavLink>
-  );
-};
+) => (
+  <NavLink
+    ref={ref}
+    className={({ isActive }) =>
+      ` ${defaultClassName} ${isActive ? activeClassName : inactiveClassName}`
+    }
+    {...props}
+  >
+    {children}
+  </NavLink>
+);
 
 export default forwardRef(CustomNavLink);
