+++ frontend/src/components/Dialogs/ConfirmLogOut/index.ts
