+++ frontend/src/components/Dialogs/index.ts
@@ -1,3 +1,4 @@
 export { default as Dialog } from './Dialog';
 export { default as ConfirmDialog } from './ConfirmDialog';
 export { default as StartTestDialog } from './StartTestDialog';
+export { default as ConfirmLogOut } from './ConfirmLogOut';
