+++ frontend/src/components/Layout/AuthorizedLayout/MobileMenu/MobileMenu.tsx
@@ -1,5 +1,4 @@
 import { FC } from 'react';
-
 import { Slide } from '@mui/material';
 
 import Avatar from 'components/Avatar';
