+++ frontend/src/components/Layout/DefaultLayout/DefaultLayout.tsx
@@ -1,18 +1,18 @@
 import { FC, Suspense } from 'react';
 import { Helmet } from 'react-helmet';
-import Grid from '@mui/material/Grid';
+import { Grid } from '@mui/material';
 
 import Loader from 'components/Loader';
-import { IDefaultLayoutProps } from 'components/Layout/types';
+import { TDefaultLayoutProps } from 'components/Layout/types';
 
-const DefaultLayout: FC<IDefaultLayoutProps> = ({ pageName, children }) => (
+const DefaultLayout: FC<TDefaultLayoutProps> = ({ pageName, children }) => (
   <>
     <Helmet>
       <title>{pageName}</title>
     </Helmet>
     <Grid container>
       <Grid item xs={12}>
-        <Suspense fallback={<Loader color="primary" />}>{children}</Suspense>
+        <Suspense fallback={<Loader />}>{children}</Suspense>
       </Grid>
     </Grid>
   </>
