+++ frontend/src/components/Routes/RoleRoute.tsx
@@ -5,7 +5,6 @@ import { useGetProfile } from 'api/profile';
 import Loader from 'components/Loader';
 import { Role } from 'constants/menuRoles';
 import { PATHS } from 'constants/routes';
-import { Loaders } from 'enums/loader';
 
 import { IRoleRouteProps } from './types';
 
@@ -13,7 +12,7 @@ const RoleRoute: FC<IRoleRouteProps> = ({ children, roles }) => {
   const { data: userProfileResponse, isLoading: isUserProfileLoading } = useGetProfile();
 
   if (isUserProfileLoading) {
-    return <Loader color="primary" type={Loaders.page} />;
+    return <Loader />;
   }
 
   const isAllowed = roles?.includes(userProfileResponse?.role as Role);
