+++ frontend/src/components/UserProfile/UserInfo/UserInfo.tsx
@@ -3,6 +3,7 @@ import { Accordion, AccordionDetails } from '@mui/material';
 import { ExpandMore as ExpandMoreIcon } from '@mui/icons-material';
 
 import { USER_INFO_LABEL } from 'constants/userInfo';
+import { convertToFullName } from 'utils/helpers/convertToFullName';
 
 import { StyledAccordionSummary, UserInfoList, UserName } from './styled';
 import UserInfoItem from './UserInfoItem';
@@ -12,7 +13,7 @@ import { IUserInfoProps } from '../types';
 
 const UserInfo: FC<IUserInfoProps> = (userInfo) => {
   const { avatar, firstName, lastName } = userInfo;
-  const userFullName = `${firstName} ${lastName}`;
+  const userFullName = convertToFullName(firstName, lastName);
   const sortedUserInfoEntries = Object.entries(userInfo)?.sort(
     ([currentInfoLabel], [nextInfoLabel]) =>
       USER_INFO_LABEL[currentInfoLabel]?.order - USER_INFO_LABEL[nextInfoLabel]?.order,
