+++ frontend/src/utils/helpers/convertToFullName.ts
@@ -1,0 +1,2 @@
+export const convertToFullName = (firstName?: string, lastName?: string) =>
+  `${firstName} ${lastName}`;
